full achitecture almost complete
