+++ client.js
@@ -1,33 +1,41 @@
-var WebSocketClient = require('websocket').client;
+const WebSocketClient = require('websocket').client;
+const client = new WebSocketClient();
+const net = require('net')
 
-var client = new WebSocketClient();
+const sendPayload = (payload) => {
+  const rpi = new net.Socket()
+  rpi.connect(2000, '10.0.0.23', () => {
+    rpi.write(JSON.stringify(payload))
+    rpi.end()
+    rpi.destroy()
+  })
+}
 
 client.on('connectFailed', function(error) {
-    console.log('Connect Error: ' + error.toString());
+  console.log('Connect Error: ' + error.toString());
 });
 
 client.on('connect', function(connection) {
-    console.log('WebSocket Client Connected');
-    connection.on('error', function(error) {
-      console.log("Connection Error: " + error.toString());
-    });
-    connection.on('close', function() {
-      console.log('echo-protocol Connection Closed');
-    });
-    connection.on('message', function(message) {
-      if (message.type === 'utf8') {
-          console.log("Received: '" + message.utf8Data + "'");
-      }
-    });
+  connection.on('error', (error) => console.log(error.toString()));
 
-    function sendNumber() {
-      if (connection.connected) {
-        var number = Math.round(Math.random() * 0xFFFFFF);
-        connection.sendUTF(number.toString());
-        setTimeout(sendNumber, 1000);
-      }
+  connection.on('close', () => console.log('echo-protocol Connection Closed'));
+
+  connection.on('message', function(message) {
+    if (message.type === 'utf8') {
+      console.log("Received: '" + message.utf8Data + "'");
     }
-    sendNumber();
+  });
+
+  const sendMsg = () => {
+    if (connection.connected) {
+      const payload = {17: true}
+      const msg = JSON.stringify(payload)
+      connection.sendUTF(msg);
+      // sendPayload(msg)
+    }
+  }
+
+  sendMsg();
 });
 
 client.connect('ws://localhost:8080/', 'rpi');
